Improve isomorphy: use the same guillotine query in the controller as at the frontend. Use the same data extraction method. Common source file for them, imported by both backend and frontend components. #48
