+++ src/main/resources/react4xp/entries/Movie.jsx
@@ -1,22 +1,5 @@
 import React from 'react'
 
-import './Movie.scss';
+import Movie from '../shared/movie/Movie';
 
-import InfoContainer from "../shared/movie/InfoContainer";
-
-const Movie = ({imageUrl, title, description, year, actors}) => (
-    <div className="movie">
-        <img className="poster"
-             src={imageUrl}
-             alt={`Movie poster: ${title}`}
-             title={`Movie poster: ${title}`}/>
-
-        <InfoContainer title={title}
-                       year={year}
-                       description={description}
-                       actors={actors}
-        />
-    </div>
-);
-
-export default Movie;
+export default (props) => <Movie {...props} />;
