+++ src/main/resources/react4xp/shared/movie/Movie.scss
