+++ src/main/resources/react4xp/shared/requestMovies.es6
@@ -1,30 +1,13 @@
-import doGuillotineRequest from "../../guillotine/guillotineRequest";
+// Frontend use by react4xp/entries/MovieList.jsx
+
+import doGuillotineRequest from "../../lib/guillotine/guillotineRequest";
+import { getListMoviesQuery, extractMovieArray } from "../../lib/movie-listing";
 
 const requestMovies = ({ apiurl, contentid, first, offset, sort, movietype, handleDataFunc }) => {
     doGuillotineRequest({
         url: apiurl,
 
-        query: `
-            query($contentid:ID!, $first:Int!, $offset:Int!, $sort:String!) {
-                guillotine {
-                    getChildren(key: $contentid, first: $first, offset: $offset, sort: $sort) {
-                        ... on ${movietype} {
-                            _id,
-                            displayName
-                            data {
-                                year
-                                description
-                                actor
-                                image {
-                                    ... on media_Image {
-                                        imageUrl(type:absolute scale:"width(300)")
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }`,
+        query: getListMoviesQuery(movietype),
 
         variables: {
             contentid,
@@ -33,18 +16,7 @@ const requestMovies = ({ apiurl, contentid, first, offset, sort, movietype, hand
             sort
         },
 
-        extractDataFunc: responseData => responseData.data.guillotine.getChildren
-            .filter( movieItem => movieItem && typeof movieItem === 'object' && Object.keys(movieItem).indexOf('data') !== -1)
-            .map(
-                movieItem => ({
-                    id: movieItem._id,
-                    title: movieItem.displayName,
-                    imageUrl: movieItem.data.image.imageUrl,
-                    year: movieItem.data.year,
-                    description: movieItem.data.description,
-                    actors: movieItem.data.actor
-                })
-            ),
+        extractDataFunc: extractMovieArray,
 
         handleDataFunc
     });
