+++ src/main/resources/site/parts/movie-list/movie-list.es6
@@ -1,44 +1,32 @@
-const contentLib = require('/lib/xp/content');
 const portal = require('/lib/xp/portal');
-const util = require('/lib/util');
 const React4xp = require('/lib/enonic/react4xp');
+
+const guillotine = require('/lib/guillotine/api')
+const { getListMoviesQuery, extractMovieArray } = require('/lib/movie-listing');
 
 
 exports.get = function(request) {
     const content = portal.getContent();
-
     const component = portal.getComponent();
 
     const sortExpression = `${component.config.sortBy} ${
         component.config.descending ? 'DESC' : 'ASC'
     }`;
 
-    const result = contentLib.getChildren({
-        key: content._path,
-        start: 0,
-        count: component.config.movieCount,
+    const movieType = `${app.name.replace(/\./g, '_')}_Movie`; // --> "com_enonic_app_react4xp_Movie" or similar
+
+    const query = getListMoviesQuery(movieType);
+
+    const variables = {
+        contentid: content._id,
+        first: component.config.movieCount,
+        offset: 0,
         sort: sortExpression
-    });
+    };
 
-    const hits = result.total > 0
-        ? util.data.forceArray(result.hits.filter( hit => hit.type === `${app.name}:movie`))
-        : [];
+    const guillotineResult = guillotine.executeQuery(query, variables);
 
-    const movies = hits.map(hit => ({
-        id: hit._id,
-        title: hit.displayName,
-        imageUrl: hit.data.image ?
-            portal.imageUrl({
-                id: hit.data.image,
-                scale: 'width(300)'
-            }) :
-            undefined,
-        description: portal.processHtml({value: hit.data.description }),
-        year: hit.data.year,
-        actors:util.data.forceArray( hit.data.actor )
-            .map( actor => (actor || '').trim())
-            .filter(actor => !!actor)
-    }));
+    const movies = extractMovieArray(guillotineResult);
 
     return React4xp.render(
         'MovieList',
@@ -47,7 +35,7 @@ exports.get = function(request) {
             apiUrl: `./${content._name}/api/guillotine`,
             parentId: content._id,
             movieCount: component.config.movieCount,
-            movieType: `${app.name.replace(/\./g, '_')}_Movie`,
+            movieType,
             sortExpression
 
         },
